+++ packages/gatsby-transformer-remark/src/__tests__/extend-node.js
@@ -1,26 +1,13 @@
-const {
-  graphql,
-  GraphQLObjectType,
-  GraphQLList,
-  GraphQLSchema,
-} = require(`gatsby/graphql`)
-const { onCreateNode, setFieldsOnGraphQLNodeType } = require(`../gatsby-node`)
-const {
-  inferObjectStructureFromNodes,
-} = require(`../../../gatsby/src/schema/infer-graphql-type`)
+const { graphql } = require(`gatsby/graphql`)
+const { onCreateNode } = require(`../gatsby-node`)
 const extendNodeType = require(`../extend-node-type`)
 
 // given a set of nodes and a query, return the result of the query
 async function queryResult(
   nodes,
   fragment,
-  { typeName = `MarkdownRemark`, types = [] } = {},
   { additionalParameters = {}, pluginOptions = {} }
 ) {
-  const inferredFields = inferObjectStructureFromNodes({
-    nodes,
-    types: [...types],
-  })
   const extendNodeTypeFields = await extendNodeType(
     {
       type: { name: `MarkdownRemark` },
@@ -37,32 +24,30 @@ async function queryResult(
     }
   )
 
-  const markdownRemarkFields = {
-    ...inferredFields,
-    ...extendNodeTypeFields,
-  }
+  const {
+    createSchemaComposer,
+  } = require(`../../../gatsby/src/schema/schema-composer`)
 
-  const schema = new GraphQLSchema({
-    query: new GraphQLObjectType({
-      name: `RootQueryType`,
-      fields: () => {
-        return {
-          listNode: {
-            name: `LISTNODE`,
-            type: new GraphQLList(
-              new GraphQLObjectType({
-                name: typeName,
-                fields: markdownRemarkFields,
-              })
-            ),
-            resolve() {
-              return nodes
-            },
-          },
-        }
-      },
-    }),
+  const {
+    addInferredFields,
+  } = require(`../../../gatsby/src/schema/infer/add-inferred-fields`)
+  const {
+    getExampleValue,
+  } = require(`../../../gatsby/src/schema/infer/example-value`)
+
+  const typeName = `MarkdownRemark`
+  const sc = createSchemaComposer()
+  const tc = sc.createTC(typeName)
+  addInferredFields({
+    schemaComposer: sc,
+    typeComposer: tc,
+    exampleValue: getExampleValue({ nodes, typeName }),
   })
+  tc.addFields(extendNodeTypeFields)
+  sc.Query.addFields({
+    listNode: { type: [tc], resolve: () => nodes },
+  })
+  const schema = sc.buildSchema()
 
   const result = await graphql(
     schema,
@@ -97,14 +82,10 @@ const bootstrapTest = (
   it(label, async done => {
     node.content = content
     const createNode = markdownNode => {
-      queryResult(
-        [markdownNode],
-        query,
-        {
-          types: [{ name: `MarkdownRemark` }],
-        },
-        { additionalParameters, pluginOptions }
-      ).then(result => {
+      queryResult([markdownNode], query, {
+        additionalParameters,
+        pluginOptions,
+      }).then(result => {
         try {
           test(result.data.listNode[0])
           done()
@@ -149,13 +130,19 @@ Where oh where is my little pony?`,
       expect(node.excerptAst).toMatchObject({
         children: [
           {
-            type: `text`,
-            value: `Where oh where is my little pony?`,
+            children: [
+              {
+                type: `text`,
+                value: `Where oh where is my little pony?`,
+              },
+            ],
+            properties: {},
+            tagName: `p`,
+            type: `element`,
           },
         ],
-        properties: {},
-        tagName: `p`,
-        type: `element`,
+        data: { quirksMode: false },
+        type: `root`,
       })
     }
   )
@@ -177,9 +164,7 @@ date: "2017-09-18T23:19:51.246Z"
       expect(node.excerpt).toMatch(``)
       expect(node.excerptAst).toMatchObject({
         children: [],
-        data: {
-          quirksMode: false,
-        },
+        data: { quirksMode: false },
         type: `root`,
       })
     }
@@ -224,9 +209,7 @@ In quis lectus sed eros efficitur luctus. Morbi tempor, nisl eget feugiat tincid
             value: `\n`,
           },
         ],
-        data: {
-          quirksMode: false,
-        },
+        data: { quirksMode: false },
         type: `root`,
       })
     },
@@ -254,7 +237,8 @@ In quis lectus sed eros efficitur luctus. Morbi tempor, nisl eget feugiat tincid
       expect(node).toMatchSnapshot()
       expect(node.excerpt.length).toBe(139)
       expect(node.excerptAst.children.length).toBe(1)
-      expect(node.excerptAst.children[0].value.length).toBe(139)
+      expect(node.excerptAst.children[0].children.length).toBe(1)
+      expect(node.excerptAst.children[0].children[0].value.length).toBe(139)
     }
   )
 
@@ -271,7 +255,8 @@ In quis lectus sed eros efficitur luctus. Morbi tempor, nisl eget feugiat tincid
       expect(node).toMatchSnapshot()
       expect(node.excerpt.length).toBe(46)
       expect(node.excerptAst.children.length).toBe(1)
-      expect(node.excerptAst.children[0].value.length).toBe(46)
+      expect(node.excerptAst.children[0].children.length).toBe(1)
+      expect(node.excerptAst.children[0].children[0].value.length).toBe(46)
     }
   )
 
@@ -288,7 +273,8 @@ In quis lectus sed eros efficitur luctus. Morbi tempor, nisl eget feugiat tincid
       expect(node).toMatchSnapshot()
       expect(node.excerpt.length).toBe(50)
       expect(node.excerptAst.children.length).toBe(1)
-      expect(node.excerptAst.children[0].value.length).toBe(50)
+      expect(node.excerptAst.children[0].children.length).toBe(1)
+      expect(node.excerptAst.children[0].children[0].value.length).toBe(50)
     }
   )
 
@@ -314,72 +300,78 @@ Where oh [*where*](nick.com) **_is_** ![that pony](pony.png)?`,
       expect(node.excerptAst).toMatchObject({
         children: [
           {
-            type: `text`,
-            value: `Where oh `,
-          },
-          {
             children: [
+              {
+                type: `text`,
+                value: `Where oh `,
+              },
               {
                 children: [
                   {
-                    type: `text`,
-                    value: `where`,
+                    children: [
+                      {
+                        type: `text`,
+                        value: `where`,
+                      },
+                    ],
+                    properties: {},
+                    tagName: `em`,
+                    type: `element`,
                   },
                 ],
-                properties: {},
-                tagName: `em`,
+                properties: {
+                  href: `nick.com`,
+                },
+                tagName: `a`,
                 type: `element`,
               },
-            ],
-            properties: {
-              href: `nick.com`,
-            },
-            tagName: `a`,
-            type: `element`,
-          },
-          {
-            type: `text`,
-            value: ` `,
-          },
-          {
-            children: [
+              {
+                type: `text`,
+                value: ` `,
+              },
               {
                 children: [
                   {
-                    type: `text`,
-                    value: `is`,
+                    children: [
+                      {
+                        type: `text`,
+                        value: `is`,
+                      },
+                    ],
+                    properties: {},
+                    tagName: `em`,
+                    type: `element`,
                   },
                 ],
                 properties: {},
-                tagName: `em`,
+                tagName: `strong`,
                 type: `element`,
+              },
+              {
+                type: `text`,
+                value: ` `,
+              },
+              {
+                children: [],
+                properties: {
+                  alt: `that pony`,
+                  src: `pony.png`,
+                },
+                tagName: `img`,
+                type: `element`,
+              },
+              {
+                type: `text`,
+                value: `?`,
               },
             ],
             properties: {},
-            tagName: `strong`,
+            tagName: `p`,
             type: `element`,
-          },
-          {
-            type: `text`,
-            value: ` `,
-          },
-          {
-            children: [],
-            properties: {
-              alt: `that pony`,
-              src: `pony.png`,
-            },
-            tagName: `img`,
-            type: `element`,
-          },
-          {
-            type: `text`,
-            value: `?`,
           },
         ],
-        properties: {},
-        tagName: `p`,
-        type: `element`,
+        data: { quirksMode: false },
+        type: `root`,
       })
     }
   )
@@ -406,28 +398,34 @@ Where is my <code>pony</code> named leo?`,
       expect(node.excerptAst).toMatchObject({
         children: [
           {
-            type: `text`,
-            value: `Where is my `,
-          },
-          {
             children: [
               {
                 type: `text`,
-                value: `pony`,
+                value: `Where is my `,
+              },
+              {
+                children: [
+                  {
+                    type: `text`,
+                    value: `pony`,
+                  },
+                ],
+                properties: {},
+                tagName: `code`,
+                type: `element`,
+              },
+              {
+                type: `text`,
+                value: ` named leo?`,
               },
             ],
             properties: {},
-            tagName: `code`,
+            tagName: `p`,
             type: `element`,
           },
-          {
-            type: `text`,
-            value: ` named leo?`,
-          },
         ],
-        properties: {},
-        tagName: `p`,
-        type: `element`,
+        data: { quirksMode: false },
+        type: `root`,
       })
     },
     { pluginOptions: { excerpt_separator: `<!-- end -->` } }
@@ -455,13 +453,19 @@ Where oh where is that pony? Is he in the stable or down by the stream?`,
       expect(node.excerptAst).toMatchObject({
         children: [
           {
-            type: `text`,
-            value: `Where oh where is that pony? Is he in the stable…`,
+            children: [
+              {
+                type: `text`,
+                value: `Where oh where is that pony? Is he in the stable…`,
+              },
+            ],
+            properties: {},
+            tagName: `p`,
+            type: `element`,
           },
         ],
-        properties: {},
-        tagName: `p`,
-        type: `element`,
+        data: { quirksMode: false },
+        type: `root`,
       })
     }
   )
@@ -522,9 +526,7 @@ Lorem ipsum dolor sit amet, consectetur adipiscing elit. Morbi auctor sit amet v
             value: `\n`,
           },
         ],
-        data: {
-          quirksMode: false,
-        },
+        data: { quirksMode: false },
         type: `root`,
       })
     },
@@ -834,48 +836,4 @@ describe(`Headings are generated correctly from schema`, () => {
       ])
     }
   )
-})
-
-describe(`Adding fields to the GraphQL schema`, () => {
-  it(`only adds fields when the GraphQL type matches the provided type`, async () => {
-    const getNode = jest.fn()
-    const getNodesByType = jest.fn()
-
-    expect(
-      setFieldsOnGraphQLNodeType({
-        type: { name: `MarkdownRemark` },
-        getNode,
-        getNodesByType,
-      })
-    ).toBeInstanceOf(Promise)
-
-    expect(
-      setFieldsOnGraphQLNodeType(
-        { type: { name: `MarkdownRemark` }, getNode, getNodesByType },
-        { type: `MarkdownRemark` }
-      )
-    ).toBeInstanceOf(Promise)
-
-    expect(
-      setFieldsOnGraphQLNodeType(
-        { type: { name: `MarkdownRemark` }, getNode, getNodesByType },
-        { type: `GatsbyTestType` }
-      )
-    ).toEqual({})
-
-    expect(
-      setFieldsOnGraphQLNodeType(
-        { type: { name: `GatsbyTestType` }, getNode, getNodesByType },
-        { type: `GatsbyTestType` }
-      )
-    ).toBeInstanceOf(Promise)
-
-    expect(
-      setFieldsOnGraphQLNodeType({
-        type: { name: `GatsbyTestType` },
-        getNode,
-        getNodesByType,
-      })
-    ).toEqual({})
-  })
 })