+++ packages/gatsby-transformer-remark/src/__tests__/on-node-create.js
@@ -1,16 +1,7 @@
+const Promise = require(`bluebird`)
 const _ = require(`lodash`)
-
 const onCreateNode = require(`../on-node-create`)
-
-const {
-  graphql,
-  GraphQLObjectType,
-  GraphQLList,
-  GraphQLSchema,
-} = require(`gatsby/graphql`)
-const {
-  inferObjectStructureFromNodes,
-} = require(`../../../gatsby/src/schema/infer-graphql-type`)
+const { graphql } = require(`gatsby/graphql`)
 
 let node
 let actions
@@ -119,79 +110,33 @@ yadda yadda
 
       expect(parsed.frontmatter.date).toEqual(new Date(date).toJSON())
     })
-
-    it(`Filters nodes with the given filter function, if provided`, async () => {
-      const content = ``
-
-      node.content = content
-      node.sourceInstanceName = `gatsby-test-source`
-
-      const createNode = jest.fn()
-      const createParentChildLink = jest.fn()
-      const actions = { createNode, createParentChildLink }
-      const createNodeId = jest.fn()
-      createNodeId.mockReturnValue(`uuid-from-gatsby`)
-
-      await onCreateNode(
-        {
-          node,
-          loadNodeContent,
-          actions,
-          createNodeId,
-        },
-        {
-          filter: node =>
-            node.sourceInstanceName === `gatsby-other-test-source`,
-        }
-      ).then(() => {
-        expect(createNode).toHaveBeenCalledTimes(0)
-        expect(createParentChildLink).toHaveBeenCalledTimes(0)
-      })
-
-      await onCreateNode(
-        {
-          node,
-          loadNodeContent,
-          actions,
-          createNodeId,
-        },
-        {
-          filter: node => node.sourceInstanceName === `gatsby-test-source`,
-        }
-      ).then(() => {
-        expect(createNode).toHaveBeenCalledTimes(1)
-        expect(createParentChildLink).toHaveBeenCalledTimes(1)
-      })
-    })
   })
 
   describe(`process graphql correctly`, () => {
     // given a set of nodes and a query, return the result of the query
-    async function queryResult(nodes, fragment, { types = [] } = {}) {
-      const schema = new GraphQLSchema({
-        query: new GraphQLObjectType({
-          name: `RootQueryType`,
-          fields: () => {
-            return {
-              listNode: {
-                name: `LISTNODE`,
-                type: new GraphQLList(
-                  new GraphQLObjectType({
-                    name: `MarkdownRemark`,
-                    fields: inferObjectStructureFromNodes({
-                      nodes,
-                      types: [...types],
-                    }),
-                  })
-                ),
-                resolve() {
-                  return nodes
-                },
-              },
-            }
-          },
-        }),
+    async function queryResult(nodes, fragment) {
+      const {
+        createSchemaComposer,
+      } = require(`../../../gatsby/src/schema/schema-composer`)
+      const {
+        addInferredFields,
+      } = require(`../../../gatsby/src/schema/infer/add-inferred-fields`)
+      const {
+        getExampleValue,
+      } = require(`../../../gatsby/src/schema/infer/example-value`)
+
+      const sc = createSchemaComposer()
+      const typeName = `MarkdownRemark`
+      const tc = sc.createTC(typeName)
+      addInferredFields({
+        schemaComposer: sc,
+        typeComposer: tc,
+        exampleValue: getExampleValue({ nodes, typeName }),
       })
+      sc.Query.addFields({
+        listNode: { type: [tc], resolve: () => nodes },
+      })
+      const schema = sc.buildSchema()
 
       const result = await graphql(
         schema,
@@ -234,8 +179,7 @@ Sed bibendum sem iaculis, pellentesque leo sed, imperdiet ante. Sed consequat ma
                     frontmatter {
                         title
                     }
-                `,
-          { types: [{ name: `MarkdownRemark` }] }
+                `
         ).then(result => {
           try {
             createdNode = result.data.listNode[0]
@@ -295,8 +239,7 @@ Sed bibendum sem iaculis, pellentesque leo sed, imperdiet ante. Sed consequat ma
                     frontmatter {
                         title
                     }
-                `,
-          { types: [{ name: `MarkdownRemark` }] }
+                `
         ).then(result => {
           try {
             createdNode = result.data.listNode[0]